Refactor parsing to be owned by processors
