+++ elyra/pipeline/__init__.py
@@ -19,7 +19,7 @@ from .handlers import PipelineExportHandler, PipelineSchedulerHandler, PipelineC
 from .parser import PipelineParser
 from .pipeline import Operation, Pipeline
 from .processor import PipelineProcessorRegistry, PipelineProcessorManager, PipelineProcessor, \
-    PipelineProcessorResponse, RuntimePipelineProcess
+    PipelineProcessorResponse, RuntimePipelineProcessor
 from .processor_kfp import KfpPipelineProcessor
 from .processor_local import LocalPipelineProcessor
 from .processor_airflow import AirflowPipelineProcessor
