+++ elyra/pipeline/processor_kfp.py
@@ -14,18 +14,19 @@
 # limitations under the License.
 #
 import autopep8
+import jupyter_core.paths
+import json
 import os
 import re
 import tempfile
 import time
 import requests
-import json
 
 from black import format_str, FileMode
 from datetime import datetime
 from elyra._version import __version__
 from elyra.metadata import MetadataManager
-from elyra.pipeline import RuntimePipelineProcess, PipelineProcessor, PipelineProcessorResponse, Operation
+from elyra.pipeline import RuntimePipelineProcessor, PipelineProcessor, PipelineProcessorResponse, Operation
 from elyra.util.path import get_absolute_path
 from jinja2 import Environment, PackageLoader
 from kfp import Client as ArgoClient
@@ -38,8 +39,12 @@ from kfp_server_api.exceptions import ApiException
 from typing import Dict
 from urllib3.exceptions import LocationValueError, MaxRetryError
 
+from .component import ComponentParser
+from .component_parser_kfp import KfpComponentParser
+from .component_registry import ComponentRegistry
 
-class KfpPipelineProcessor(RuntimePipelineProcess):
+
+class KfpPipelineProcessor(RuntimePipelineProcessor):
     _type = 'kfp'
 
     # Provide users with the ability to identify a writable directory in the
@@ -51,6 +56,29 @@ class KfpPipelineProcessor(RuntimePipelineProcess):
     @property
     def type(self):
         return self._type
+
+    @property
+    def component_catalog(self) -> str:
+        return self._component_catalog_location
+
+    @property
+    def component_parser(self) -> ComponentParser:
+        return self._component_parser
+
+    def __init__(self, root_dir, **kwargs):
+        super().__init__(root_dir, **kwargs)
+        # then sys.prefix, where installed files will reside (factory data)
+        self._component_catalog_location =  \
+            os.path.join(jupyter_core.paths.ENV_JUPYTER_PATH[0],
+                         'components',
+                         f"{self._type}_component_catalog.json")
+
+        if not os.path.exists(self._component_catalog_location):
+            raise FileNotFoundError(f'Invalid component catalog path {self._component_catalog_location}'
+                                    f' for {self._type} processor')
+
+        self._component_parser = KfpComponentParser()
+        self._component_registry = ComponentRegistry(self.component_catalog, self.component_parser)
 
     def process(self, pipeline):
         """Runs a pipeline on Kubeflow Pipelines
