+++ packages/pipeline-editor/src/PipelineService.tsx
@@ -86,10 +86,10 @@ export class PipelineService {
    */
   static async getComponentProperties(
     runtimeName: string,
-    componentCategory: string
+    componentId: string
   ): Promise<any> {
     return RequestHandler.makeGetRequest(
-      `elyra/pipeline/components/${runtimeName}/${componentCategory}/properties`
+      `elyra/pipeline/components/${runtimeName}/${componentId}/properties`
     );
   }
